requirements.txt
selenium
webdriver-manager
pandas
supabase
python-dotenv
